version: Always prefer builtin importlib.metadata
If we have a builtin importlib.metadata (Python 3.8+) and the importlib_metadata
backport installed, we preferred the backport. However, the version.py tests do
the opposite: They only mock the builtin if it is available. This did lead to
failing tests if the backport was installed in an environment where the builtin
was available too.

Since we don't need any specialized functionality (only reading the version), we
can prefer the builtin no matter whether a backport is available or not.